refactor: Optimize code for generating post comments and users
This commit refactors the code in the `makePosts` function to optimize the process of generating post comments and users. It introduces a more efficient approach by fetching the necessary data in bulk using SQL queries instead of making multiple individual queries. This improves the performance and reduces the number of database calls, resulting in faster response times.

The changes include:
- Consolidating the retrieval of post comments and users into single SQL queries
- Storing the retrieved data in maps for efficient lookup
- Limiting the number of comments displayed to 3, unless specified otherwise

These optimizations enhance the overall efficiency and scalability of the application.